src/main/java/quotify_app/usecases/login/LoginInteractor.java
package quotify_app.usecases.login;

import quotify_app.entities.User;

/**
 * The Login Interactor.
 */
public class LoginInteractor implements LoginInputBoundary {
    private final LoginUserDataAccessInterface userDataAccessObject;
    private final LoginOutputBoundary loginPresenter;

    public LoginInteractor(LoginUserDataAccessInterface userDataAccessInterface,
                           LoginOutputBoundary loginOutputBoundary) {
        this.userDataAccessObject = userDataAccessInterface;
        this.loginPresenter = loginOutputBoundary;
    }

    @Override
    public void execute(LoginInputData loginInputData) {
        final String username = loginInputData.getUsername();
        final String password = loginInputData.getPassword();

        if (!userDataAccessObject.existsByName(username)) {
            loginPresenter.prepareFailView(username + ": Account does not exist.");
        }
        else {
            final User user = userDataAccessObject.get(username);
            if (!password.equals(user.getPassword())) {
                loginPresenter.prepareFailView("Incorrect password for \"" + username + "\".");
            }
            else {
                final LoginOutputData loginOutputData = new LoginOutputData(user.getName());
                loginPresenter.prepareSuccessView(loginOutputData);
                goToFunction();
            }
        }
    }

    /**
     * Trigger view transition to Signup View through the presenter.
     */
    @Override
    public void goToSignup() {
        loginPresenter.goToSignup();
    }

    /**
     * Trigger view transition to Function View through the presenter.
     */
    @Override
    public void goToFunction() {
        loginPresenter.goToFunctionView();
    }
}
